marsRover13 Separates middleware and business logic
