+++ server/services/getPhotoOfTheDay.js
@@ -1,0 +1,24 @@
+const axios = require('axios');
+const { NASA_API_KEY } = require('../../config/apiCredentials');
+
+const queryParameters = {
+	params: {
+		api_key: NASA_API_KEY,
+	},
+};
+
+const getPhotoOfTheDay = async () => {
+	try {
+		const apodData = await axios.get(
+			'https://api.nasa.gov/planetary/apod',
+			queryParameters
+		);
+		return apodData;
+	} catch (error) {
+		return Promise.reject(new Error(`${error.name}: ${error.message}`));
+	}
+};
+
+module.exports = {
+	getPhotoOfTheDay,
+};
